+++ app/containers/add-chores.js
@@ -1,0 +1,12 @@
+import React from 'react';
+// import { withRouter } from 'react-router';
+import Menu from '../components/menu';
+
+const AddChores = () => (
+  <div>
+    <Menu />
+    <h2>AddChores</h2>
+  </div>
+);
+
+export default AddChores;
