+++ app/containers/home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Menu from '../components/menu';
+
+const Home = () => (
+  <div>
+    <Menu />
+    <h2>Home ixi</h2>
+  </div>
+);
+
+export default Home;
