--- app/home.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Menu from './components/menu';
-
-const Home = () => (
-  <Menu />
-);
-
-export default Home;
